WhichFish
Which fish is at its most delicious in which months? In which months is it at least edible? What are the cooking methods? What are the cooking recipes? I started to develop this application in order to find answers to these questions and not be undecided in the fish aisleway.